Move closure checking into its own file. Shrink check/mod.rs!
